Consolidate linting to Ruff, remove Black/isort/Flake8
- Remove black and isort from dev dependencies (poetry.lock updated)
- Delete [tool.black] and [tool.isort] config sections from pyproject.toml
- Replace 3 pre-commit hooks (black, isort, flake8) with 2 ruff hooks (format, check)
- Update CI/CD workflow to use `ruff format --check` and `ruff check`
- Consolidate Makefile targets to ruff-only commands
- Ruff now handles: formatting (Black-compatible), import sorting (isort rules), and linting (Flake8 rules)

+++ app/strategies/mean_reversion/tools/signal_generation.py
@@ -164,7 +164,7 @@ def generate_current_signals(config: Config, log: Callable) -> pl.DataFrame:
         step_pct = config.get("CHANGE_PCT_STEP", 0.01)
 
         # Calculate number of steps to ensure we include end_pct
-        num_steps = int(round((end_pct - start_pct) / step_pct)) + 1
+        num_steps = round((end_pct - start_pct) / step_pct) + 1
 
         # Create parameter arrays with controlled precision
         change_pcts = [round(start_pct + i * step_pct, 2) for i in range(num_steps)]
